API fixes according to new guideline

+++ packages/alea-frontend/pages/api/job-portal/delete-job-post.ts
@@ -2,17 +2,13 @@ import { Action, CURRENT_TERM, ResourceName } from '@stex-react/utils';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { getUserIdIfAuthorizedOrSetError } from '../access-control/resource-utils';
 import { checkIfPostOrSetError, executeAndEndSet500OnError } from '../comment-utils';
-import { getJobCategoryUsingIdOrSetError } from './update-job-type';
-import { getJobPostUsingIdOrSetError } from './update-job-post';
+import { getJobCategoryUsingIdOrSet500OnError } from './update-job-type';
+import { getJobPostUsingIdOrSet500OnError } from './update-job-post';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (!checkIfPostOrSetError(req, res)) return;
   const { id } = req.body;
-  if (!id) return res.status(400).send('Job Post id is missing');
-
-  const currentJobPost = await getJobPostUsingIdOrSetError(id, res);
-  if (!currentJobPost) return;
-
+  if (!id) return res.status(422).send('Job Post id is missing');
   const userId = await getUserIdIfAuthorizedOrSetError(
     req,
     res,
@@ -22,7 +18,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
   );
   if (!userId) return;
 
-
+  const currentJobPost = await getJobPostUsingIdOrSet500OnError(id, res);
+  if (!currentJobPost) return;
   const result = await executeAndEndSet500OnError('DELETE FROM jobpost WHERE id = ?', [id], res);
   if (!result) return;
   res.status(200).end();
